backend-workout/requirements.txt
fastapi
uvicorn
langchain
langchain_huggingface
chromadb
pymupdf
python-docx
pypdf
ollama
requests
langchain_chroma
langchain_ollama